consider landscape for resizing images: https://github.com/metanorma/metanorma-iho/issues/164
