+++ lib/html2doc.rb
@@ -4,3 +4,4 @@ require_relative "html2doc/mime"
 require_relative "html2doc/notes"
 require_relative "html2doc/math"
 require_relative "html2doc/lists"
+require_relative "html2doc/xml"
