+++ lib/html2doc/base.rb
@@ -30,8 +30,7 @@ class Html2Doc
   end
 
   def process_header(headerfile)
-    return if headerfile.nil?
-
+    headerfile.nil? and return
     doc = File.read(headerfile, encoding: "utf-8")
     doc = header_image_cleanup(doc, @dir1, @filename,
                                File.dirname(@filename))
@@ -66,6 +65,7 @@ class Html2Doc
   end
 
   def cleanup(docxml)
+    locate_landscape(docxml)
     namespace(docxml.root)
     image_cleanup(docxml, @dir1, @imagedir)
     mathml_to_ooml(docxml)
@@ -76,76 +76,11 @@ class Html2Doc
     docxml
   end
 
-  NOKOHEAD = <<~HERE.freeze
-    <!DOCTYPE html SYSTEM
-    "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
-    <html xmlns="http://www.w3.org/1999/xhtml">
-    <head> <title></title> <meta charset="UTF-8" /> </head>
-    <body> </body> </html>
-  HERE
-
-  def to_xhtml(xml)
-    xml.gsub!(/<\?xml[^>]*>/, "")
-    unless /<!DOCTYPE /.match? xml
-      xml = '<!DOCTYPE html SYSTEM
-          "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">' + xml
-    end
-    xml = xml.gsub(/<!--\s*\[([^\]]+)\]>/, "<!-- MSWORD-COMMENT \\1 -->")
-      .gsub(/<!\s*\[endif\]\s*-->/, "<!-- MSWORD-COMMENT-END -->")
-    Nokogiri::XML.parse(xml)
+  def locate_landscape(_docxml)
+    css = read_stylesheet(@stylesheet)
+    @landscape = css.scan(/div\.\S+\s+\{\s*page:\s*[^;]+L;\s*\}/m)
+      .map { |e| e.sub(/^div\.(\S+).*$/m, "\\1") }
   end
-
-  DOCTYPE = <<~"DOCTYPE".freeze
-    <!DOCTYPE html SYSTEM "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
-  DOCTYPE
-
-  def from_xhtml(xml)
-    xml.to_xml.sub(%{ xmlns="http://www.w3.org/1999/xhtml"}, "")
-      .sub(DOCTYPE, "").gsub(%{ />}, "/>")
-      .gsub(/<!-- MSWORD-COMMENT (.+?) -->/, "<!--[\\1]>")
-      .gsub(/<!-- MSWORD-COMMENT-END -->/, "<![endif]-->")
-      .gsub("\n--&gt;\n", "\n-->\n")
-  end
-
-  def msword_fix(doc)
-    # brain damage in MSWord parser
-    doc.gsub!(%r{<w:DoNotOptimizeForBrowser></w:DoNotOptimizeForBrowser>},
-              "<w:DoNotOptimizeForBrowser/>")
-    doc.gsub!(%r{<span style="mso-special-character:footnote"/>},
-              '<span style="mso-special-character:footnote"></span>')
-    doc.gsub!(%r{<div style="mso-element:footnote-list"></div>},
-              '<div style="mso-element:footnote-list"/>')
-    doc.gsub!(%r{(<a style="mso-comment-reference:[^>/]+)/>}, "\\1></a>")
-    doc.gsub!(%r{<link rel="File-List"}, "<link rel=File-List")
-    doc.gsub!(%r{<meta http-equiv="Content-Type"},
-              "<meta http-equiv=Content-Type")
-    doc.gsub!(%r{></m:jc>}, "/>")
-    doc.gsub!(%r{></v:stroke>}, "/>")
-    doc.gsub!(%r{></v:f>}, "/>")
-    doc.gsub!(%r{></v:path>}, "/>")
-    doc.gsub!(%r{></o:lock>}, "/>")
-    doc.gsub!(%r{></v:imagedata>}, "/>")
-    doc.gsub!(%r{></w:wrap>}, "/>")
-    doc.gsub!(%r{<(/)?m:(span|em)\b}, "<\\1\\2")
-    doc.gsub!(%r{&tab;|&amp;tab;},
-              '<span style="mso-tab-count:1">&#xA0; </span>')
-    doc.split(%r{(<m:oMath>|</m:oMath>)}).each_slice(4).map do |a|
-      a.size > 2 and a[2] = a[2].gsub(/>\s+</, "><")
-      a
-    end.join
-  end
-
-  PRINT_VIEW = <<~XML.freeze
-
-    <xml>
-    <w:WordDocument>
-    <w:View>Print</w:View>
-    <w:Zoom>100</w:Zoom>
-    <w:DoNotOptimizeForBrowser/>
-    </w:WordDocument>
-    </xml>
-    <meta http-equiv='Content-Type' content="text/html; charset=utf-8"/>
-  XML
 
   def define_head1(docxml, _dir)
     docxml.xpath("//*[local-name() = 'head']").each do |h|
@@ -174,7 +109,6 @@ class Html2Doc
     # xml.children.first << Nokogiri::XML::Comment.new(xml, s)
     xml.children.first << Nokogiri::XML::CDATA
       .new(xml, "\n<!--\n#{stylesheet}\n-->\n")
-
     xml.root.to_s
   end
 
@@ -199,30 +133,15 @@ class Html2Doc
       head.add_child css
     elsif title.nil?
       head.children.first.add_previous_sibling css
-    else
-      title.add_next_sibling css
+    else title.add_next_sibling css
     end
-  end
-
-  def namespace(root)
-    {
-      o: "urn:schemas-microsoft-com:office:office",
-      w: "urn:schemas-microsoft-com:office:word",
-      v: "urn:schemas-microsoft-com:vml",
-      m: "http://schemas.microsoft.com/office/2004/12/omml",
-    }.each { |k, v| root.add_namespace_definition(k.to_s, v) }
-  end
-
-  def rootnamespace(root)
-    root.add_namespace(nil, "http://www.w3.org/TR/REC-html40")
   end
 
   def bookmarks(docxml)
     docxml.xpath("//*[@id][not(@name)][not(@style = 'mso-element:footnote')]")
       .each do |x|
-      next if x["id"].empty? ||
-        %w(shapetype v:shapetype shape v:shape).include?(x.name)
-
+      (x["id"].empty? ||
+        %w(shapetype v:shapetype shape v:shape).include?(x.name)) and next
       if x.children.empty? then x.add_child("<a name='#{x['id']}'></a>")
       else x.children.first.previous = "<a name='#{x['id']}'></a>"
       end
